% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_vars_funs.R
\name{function_name}
\alias{function_name}
\title{Function name
	
\code{function_name} is a function that does}
\usage{
function_name(par1)
}
\arguments{
\item{par1}{Explanation of the role of parameter 1}
}
\value{
theExplanation of what is returned (z)
}
\description{
Function name
	
\code{function_name} is a function that does
}

