% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_vars_funs.R
\docType{data}
\name{myglobal}
\alias{myglobal}
\title{Global specs for rcsurplus}
\format{An object of class \code{list} of length 9.}
\usage{
myglobal
}
\arguments{
\item{priorK}{prior for carrying capacity}

\item{priorr}{prior for intrinsic growth rate}

\item{priorq}{prior for catchability}

\item{priors}{prior for observational variance}

\item{mcmc_n}{number of Markov Chain Monte Carlo iterations}

\item{mcmc_t}{MCMC thinning}

\item{mcmc_b}{MCMC burn-in fraction}

\item{mcmc_c}{number of MCMC chains}
}
\description{
\code{rcsurplus1d.defaults} is a list that contains default 
parameters for this package
}
\keyword{datasets}

