% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\docType{class}
\name{rcplotter-class}
\alias{rcplotter}
\alias{rcplotter-class}
\title{rcplotter}
\description{
rcplotter
}
\section{Fields}{

\describe{
\item{\code{buffer}}{matrix.}

\item{\code{palettes}}{list.}

\item{\code{default_palettes}}{list.}
}}
\section{Methods}{

\describe{
\item{\code{get_buffer_plot()}}{Prints the plots in the buffer}

\item{\code{get_density_plot(out, mytitle, dolegend = TRUE, xpos = 1, ypos = 1)}}{Plots the posterior density of >= 1 hyperparameters}

\item{\code{get_one_density_plot(out, mytitle, xpos = 1, ypos = 1)}}{Plots the posterior density of one hyperparameter}

\item{\code{get_scatterplot(out, mytitle, xpos = 1, ypos = 1)}}{Plots a scatterplot of two hyperparameters}

\item{\code{get_ts_fit_plot(out, mytitle, mylabs, xpos = 1, ypos = 1)}}{Plots a time series plot with observed and fit (median plus 95 CI)}

\item{\code{initialize(name = "anonymous", verbose = TRUE, autoconstruct = FALSE,
  specs = NULL)}}{Default method to initialize basic objects}

\item{\code{set_buffer_size(nr, nc)}}{Sets up plotter's buffer size (number of rows and columns)}

\item{\code{set_in_buffer(myplot, xpos, ypos)}}{Places a plot in the buffer}

\item{\code{set_palette(argument, value)}}{Sets the printer's colour palettes}
}}

