% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplus1d.R
\docType{class}
\name{rcsurplus1d-class}
\alias{rcsurplus1d}
\alias{rcsurplus1d-class}
\title{rcsurplus}
\description{
rcsurplus
}
\section{Fields}{

\describe{
\item{\code{fit_counter}}{numeric. NUmber of times the model has been fitted.}

\item{\code{models}}{list. List of models handled by the daemon.}

\item{\code{coda_models}}{list. List of CODA output.}

\item{\code{model_names}}{character. List with names of fitted models.}

\item{\code{par_titles}}{data.frame. Parameter names, as they appear in plots.}

\item{\code{IO}}{.IAT. Reference class object that handles OpenBUGS output.}
}}
\section{Methods}{

\describe{
\item{\code{fit_models(input)}}{Function to fit surplus production models}
}}

