% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{data}
\name{rcsurplus1d.help}
\alias{rcsurplus1d.help}
\title{Global specs for rcsurplus}
\format{An object of class \code{list} of length 9.}
\usage{
rcsurplus1d.help
}
\description{
\code{rcsurplus1d.help}
}
\keyword{datasets}

