% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplus1d.R
\docType{class}
\name{rcsurplus1d-class}
\alias{rcsurplus1d}
\alias{rcsurplus1d-class}
\title{rcsurplus}
\description{
rcsurplus
}
\section{Fields}{

\describe{
\item{\code{fit_counter}}{numeric.}

\item{\code{models}}{list.}

\item{\code{coda_models}}{list.}

\item{\code{model_names}}{character.}

\item{\code{par_titles}}{data.frame.}

\item{\code{IO}}{.IAT.}
}}
\section{Methods}{

\describe{
\item{\code{fit_models(input)}}{Function to fit surplus production models}
}}

